Add file_exists guard clause on swift.upload_file function

+++ extlinks/common/swift.py
@@ -181,12 +181,15 @@ def upload_file(
     object_name = os.path.basename(path)
 
     with open(path, "rb") as f:
-        conn.put_object(
-            container,
-            object_name,
-            contents=f,
-            content_type=content_type,
-        )
+        if not file_exists(conn, container, object_name):
+            conn.put_object(
+                container,
+                object_name,
+                contents=f,
+                content_type=content_type,
+            )
+        else:
+            return False
 
     return object_name
 
@@ -290,6 +293,8 @@ def batch_upload_files(
 
             try:
                 object_name = future.result()
+                if not object_name:
+                    logger.info(f"Skipping upload for '%s' - already uploaded", path)
                 logger.info(f"Successfully uploaded '%s' as '%s'", path, object_name)
                 successful.append(path)
             except Exception as exc:
